+++ cli/src/index.ts
@@ -1,4 +1,4 @@
-import program from 'commander';
+import { program } from 'commander';
 
 import c from './colors';
 import { checkExternalConfig, loadConfig } from './config';