+++ core/CHANGELOG.md
@@ -3,10 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-beta.0](https://github.com/ionic-team/capacitor/compare/3.6.0...4.0.0-beta.0) (2022-06-17)
+
+**Note:** Version bump only for package @capacitor/core
+
+
+
+
+
 # [3.6.0](https://github.com/ionic-team/capacitor/compare/3.5.1...3.6.0) (2022-06-17)
 
 **Note:** Version bump only for package @capacitor/core
 
+
+
+
+
+# [4.0.0-alpha.2](https://github.com/ionic-team/capacitor/compare/3.4.1...4.0.0-alpha.2) (2022-05-12)
+
+**Note:** Version bump only for package @capacitor/core
 
 
 
@@ -23,6 +38,15 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [3.5.0](https://github.com/ionic-team/capacitor/compare/3.4.3...3.5.0) (2022-04-22)
+
+**Note:** Version bump only for package @capacitor/core
+
+
+
+
+
+
+# [4.0.0-alpha.1](https://github.com/ionic-team/capacitor/compare/3.4.1...4.0.0-alpha.1) (2022-03-25)
 
 **Note:** Version bump only for package @capacitor/core